Add `return_distance_threshold = 0.1` to unit tests
Unit tests for Hesai decoders were not setting
`dual_return_distance_threshold` at all, causing it to be `0`
implicitly. This caused a large number of duplicate points to be
output by the decoders.

The distance is set to `0.1 [m]` for all Hesai sensors now, which is the
default in the Hesai launch file.