Complete Hesai test refactoring into generic implementation
The sensor-specific duplicates of the ..._test_main and ..._test files
have been refactored into a single generic implementation that takes
only the sensor parameters as input.

The ReadBag function has been refactored such that it takes a callback
function which is called on every point cloud scan.
This allows for easy and concise implementation of new test cases.