refactor(robosense): move Helios, Bpearl to composable sensor implementation, rewrite decoder
In anticipation of the Robosense M1 PR, the `robosense_decoder.hpp` had to become more generic to support non-angle-based scan cutting along with completely different mechanisms for angle data, dual return and timestamps.

Thus, Robosense is the first vendor to be moved to the new composable sensor model, starting with Helios and Bpearl.

* Introduce decode groups: multiple packets can be grouped and decoded together, as is necessary for RSM1
* Packet/Block/Unit stride definition for generically handling different return group layouts
* Decoder now uses generic sensor member functions to access all point data

* Rewrite sensor classes to use SensorBase and mixins
* Add SensorInfo classes for the info decoder
* Currently, timing and return types are not fully re-implemented
* `angle_corrector.hpp` is deleted in favor of sensor mixins

* The driver now instantiates sensor objects passed to the decoder